+++ Servidor/src/entity/ConfeccionEntity.java
@@ -2,9 +2,18 @@ package entity;
 
 import java.io.Serializable;
 import java.util.ArrayList;
+import java.util.Date;
 import java.util.List;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.JoinColumn;
+import javax.persistence.OneToMany;
+import javax.persistence.OneToOne;
+import javax.persistence.Table;
 
 import negocio.Confeccion;
 import negocio.Insumo;
@@ -17,7 +26,7 @@ public class ConfeccionEntity implements Serializable{
 	@GeneratedValue(strategy=GenerationType.AUTO)
 	private int id;
 	
-	private int tiempoProd;
+	private Float tiempoProd;
 	private String detalle;
 	
 	@OneToOne()
@@ -50,11 +59,11 @@ public class ConfeccionEntity implements Serializable{
 		this.id = id;
 	}
 
-	public int getTiempoProd() {
+	public Float getTiempoProd() {
 		return tiempoProd;
 	}
 
-	public void setTiempoProd(int tiempoProd) {
+	public void setTiempoProd(Float tiempoProd) {
 		this.tiempoProd = tiempoProd;
 	}
